This adds an additional condition on my previous patch in r183 which tried to fix box characters appearing in some pages, such as Gmail.
The problem is that the newline character is treated as a zero-width space in the old patch, but apparently WebKit relies on this turning into a space (or something, I don't totally understand). I just removed control characters from this condition since we don't actually want to change their behavior in this respect.

webkit=https://bugs.webkit.org/show_bug.cgi?id=20237
BUG=1166024

git-svn-id: http://src.chromium.org/svn/trunk/src@204 4ff67af0-8c30-449e-8e8b-ad334ec8d88c