Fix crash due to race condition in ImportingProgressView.
What is happening is that the ImporterView creates an ImportingProgressView to show the progress of the import but the ImportingProgressView receives the Cancel event before ImportStarted, so importing_ is not set to true. Cancel will think import is done and returns true (allowing the view to be destroyed) and then the ImporterView will try to communicate with a dead observer.

By initializing importing_ to true this crash is avoided (we wouldn't be creating the view unless we were about to import, right?).

BUG=1320194
TEST=Open Import dialog, press Enter and then Esc really quickly and Chrome should not crash (try it 10 times to be sure).


git-svn-id: http://src.chromium.org/svn/trunk/src@702 4ff67af0-8c30-449e-8e8b-ad334ec8d88c