Dupe of http://codereview.chromium.org/238
The download page was quickly changed to address  bug 1358202, but still there's
a problem that the UI
langauge of Chrome set-up program is different from what a user choose in the
following
scenarios:

  * English is chosen on non-en-US OS
  * Hebrew is chosen on non-Hebrew OS
  * Norwegian Bokmal is selected on non-Norwegian Bokmal OS

Once these changes go in for next-point-release, the change made on the download
server can be reverted. 

1. Aliases {iw,no,tl} to {he,nb,fil}. 
2. Also map en to en-US.
3. Map {zh-HK, zh-MK, zh-SG} to {zh-TW, zh-TW, zh-CN}

This does not yet deal with other potentially problematic cases.

Added a similar change for l10n_string_util.cc for installer .

Tony already LGTM'd 238 so that I'm tbr'ing here.


BUG=1358202
TEST=L10n* unittest
TBR=tc



Review URL: http://codereview.chromium.org/260

git-svn-id: http://src.chromium.org/svn/trunk/src@1800 4ff67af0-8c30-449e-8e8b-ad334ec8d88c