Move more platform-specific stuff from WebContents to the view.
I refactored the way title setting is done since there were two parts that duplicated the logic and did so slightly differently.

I removed some old code in RendererGone whose premise is no longer valid and had the effect of creating a bug with no benefit.

I moved the sad tab view from WebContents to the view. Amusingly, the view was the only one using this, but it would go poke inside of the WebContents to get it.
Review URL: http://codereview.chromium.org/7650

git-svn-id: http://src.chromium.org/svn/trunk/src@3601 4ff67af0-8c30-449e-8e8b-ad334ec8d88c