When a windowless plugin is hanging, the browser process is also hanging because of the "dummy activation window".
Unfortunately we used to skip the hung plugin notification
dialog for this window. I'm changing this behavior to display
it even for this window.

We don't know what plugin it is, so we just display "an
unknown plugin", but this is better than nothing.

This is a potential fix for 3383.

I do not expect to completely fix 3383 with it, but it
should help minimize the problem.

bug:3383


Review URL: http://codereview.chromium.org/9386

git-svn-id: http://src.chromium.org/svn/trunk/src@4792 4ff67af0-8c30-449e-8e8b-ad334ec8d88c