Ensure that windowed plugins get focus during WM_MOUSEACTIVATE. This fixes bug http://code.google.com/p/chromium/issues/detail?id=4273, which shows up when windowed plugins enter modal loops like context menu and keyboard navigation does not work in the menu.
We were not converting the coordinates returned by GetCursorPos to client coordinates correctly, which caused us to miss the child window (Plugin) at times.

This also fixes bug http://code.google.com/p/chromium/issues/detail?id=937, which was an issue with menu selections in Java menus. This occured because we would force 
the child window to have focus in every WM_MOUSEACTIVATE message. In this case the Java plugin window, which is a child of the plugin window already has focus. It receives a WM_KILLFOCUS message due to the forced SetFocus in our handler and takes out the menu, thus ignoring the menu selection. Fix for this issue is to handle WM_MOUSEACTIVATE only if a child window of RenderWidgetHostHWND does not have focus.

R=amit
Bug=4273

Review URL: http://codereview.chromium.org/10004

git-svn-id: http://src.chromium.org/svn/trunk/src@5178 4ff67af0-8c30-449e-8e8b-ad334ec8d88c