Omnibox metrics logging patch splitout, part 2: Remove the AutocompleteMatch NULL constructor.  It's too easy to forget to set various members with this.
The changes from .resize() to .erase() are necessary because the compiler doesn't know resize() won't be enlarging the vector and thus needing to access the NULL constructor.  The changes to the HistoryContents shortcut code were similarly necessary to avoid a NULL construction, but in the end I think made the resulting code a bit clearer.
Review URL: http://codereview.chromium.org/10837

git-svn-id: http://src.chromium.org/svn/trunk/src@5271 4ff67af0-8c30-449e-8e8b-ad334ec8d88c