Omnibox metrics logging patch splitout, part 3: Pass match.type to the edit and let it make the determination about showing the search hint, instead of making the determination in the popup.
In the ultimate patch, the types on which we'll show the hint will expand, so this saved duplicating some logic at points in the popup; in retrospect it makes more sense anyway, since there's no reason for the popup to know anything about the search hint.  That's more the edit's job anyway.
Review URL: http://codereview.chromium.org/10839

git-svn-id: http://src.chromium.org/svn/trunk/src@5276 4ff67af0-8c30-449e-8e8b-ad334ec8d88c