Normalise layout test line endings to use UNIX style.
On Linux, the diffs against the Windows baselines fail if the baseline
has \r\n line endings. The output from the test shell already appears to
be normalised to \n endings, so it's not clear to me how Windows diff
works in this case unless the Windows tools are automatically detecting
and fixing the line endings.

Review URL: http://codereview.chromium.org/11280


git-svn-id: http://src.chromium.org/svn/trunk/src@5689 4ff67af0-8c30-449e-8e8b-ad334ec8d88c