Added GC protection support for objects with possible pending activity which is only backed by weak wrappers.
The objects which can have pending activity are XMLHttpRequest and MessagePort. In WebKit these two classes are both instances of ActiveDOMObject. In the V8 binding layer these two types of objects are now beeing tracked so that before GC their wrappers can be made not weak if there is currently pending activity. In addition to that MessagePort wrappers now have an additional internal field to reference their entangled port when entangled.

Before MessagePort where added to WebKit pending activity for XMLHttpRequest was handled through a direct GC protect mechanism which protected/unprotected the object when its state changed. However, using this for MessagePort would require an GC protection check in each MessagePort message as the transition to/from pending is not as explicit as for XMLHttpRequest.

Changed direct calls to dom_object_map().set(...) to use SetJSWrapperForDOMObject and added an ASSERT to check that SetJSWrapperForDOMObject is called with a DOM wrapper object. Changed the order of calls to SetDOMWrapper and SetJSWrapperForDOMObject to make sure that SetDOMWrapper is called first.

This fixes 3 GC related layout tests
  chrome/fast/dom/xmlhttprequest-gc.html
  LayoutTests/fast/events/message-channel-gc-2.html
  LayoutTests/fast/events/message-channel-gc-3.html

Added a chrome specific GC layout test for MessagePort (chrome/fast/dom/xmlhttprequest-gc.html) to match the existing one for XMLHttpRequest (chrome/fast/dom/xmlhttprequest-gc.html).

This change forks MessagePort.cpp.
Review URL: http://codereview.chromium.org/11205

git-svn-id: http://src.chromium.org/svn/trunk/src@5770 4ff67af0-8c30-449e-8e8b-ad334ec8d88c