Update to latest trunk version of V8.
This includes Evan's change to use char instead of wchar_t for counter
names.

Also, because of changes to the regexp parser, we need to rebaseline a
couple of tests mainly because of error message changes.  Most of the
tests were already rebaselined because of differences in error
messages.
Review URL: http://codereview.chromium.org/12902

git-svn-id: http://src.chromium.org/svn/trunk/src@6349 4ff67af0-8c30-449e-8e8b-ad334ec8d88c