Change init() in GKURL to use the default UTF-8 converter when an encoding is one of Unicode encodings (UTF-8,7,16,32).  HTML5 stipulates that it should use UTF-8 for UTF-16, which is extended in this change list to include  UTF-32 and UTF-7 (Firefox made that change a long time ago).
The corresponding webkit bug is http://bugs.webkit.org/show_bug.cgi?id=21635 (the patch was landed in r38755 in the webkit trunk and picked up in r6182 in our copy of Webkit).
The layout tests for this was landed in the webkit in r38988 : http://bugs.webkit.org/show_bug.cgi?id=22492 ) and will be picked up in the next webkit merge.

bug=3433
TEST=pass the layout tests http/tests/misc/{url,submit-get,submit-post}-in-utf{7,32be,32le,16le,16be}.html (not yet in our tree)


Review URL: http://codereview.chromium.org/13186

git-svn-id: http://src.chromium.org/svn/trunk/src@6481 4ff67af0-8c30-449e-8e8b-ad334ec8d88c