Misc proxy service changes.
(1) Changed the proxy service ownership model -- rather than being a
detail of the HTTP stack, it is now a dependency owned by
UrlRequestContext.
 - ProxyService is owned by UrlRequestContext (before was
   HttpNetworkSession)
 - ProxyResolver is owned by ProxyService (before was
   HttpNetworkSession)

Being able to share the proxy service is needed in several places,
including incognito mode http context (http://crbug.com/3564), and for
proxy resolving in the new FTP stack.

(2) Added an IPC for getting of the ProxyResolverWinHttp dependency in
the plugin process. Not hooked up yet, but intent is to route the
proxy resolve requests through the browser process.

(3) Changed some unit tests which were depending on the system proxy
settings (this was a sideffect of their calling
HttpNetworkLayer::CreateFactory(NULL)).

(4) Moved the first-time ProxyService::UpdateConfig out of the
constructor and into the initial request. Done to avoid startup
perf regressions, since the ProxyService construction is now done
earlier (on the startup critical path).

BUG=3564
Review URL: http://codereview.chromium.org/12938

git-svn-id: http://src.chromium.org/svn/trunk/src@6693 4ff67af0-8c30-449e-8e8b-ad334ec8d88c