Fix 'oops' in FontUtilsChromiumWin.cpp that was causing some layout tests to crash.
I failed to handle USCRIPT_INVALID_CODE as an input to GetFontFamilyForScript.
That constant has a value of -1 which caused my array indexing to be unhappy.

I also corrected the 'initialized' flag to be a static so that we would not
waste cycles calling InitailizeScriptFontMap over-and-over again.  Yikes!

R=ojan

Review URL: http://codereview.chromium.org/13722

git-svn-id: http://src.chromium.org/svn/trunk/src@6744 4ff67af0-8c30-449e-8e8b-ad334ec8d88c