* Launch Chrome in non-elevated mode.
Currently Google Update LaunchCmdLine launches command with the callers token. To launch non-elevated we try to get the token of explorer.exe and impersonate it (same as what Google Update does in another place).

Also fix a bug in Compatibility check function. We should exit if we fail to load advapi32.dll and not if we succeed.

I am still doing testing with various builds of Google Update.

Review URL: http://codereview.chromium.org/12970

git-svn-id: http://src.chromium.org/svn/trunk/src@7023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c