Fix for crbug.com/5471 (browser crash regression).
The "pending_save_manager_" used to only be non-null
between the time the user hits 'Submit' on a form and we either 1) show an 
infobar, or 2) update an existing login. Now, instead, the member only gets reset on 
InfoBarClosed(), and so this ordering of events was causing a crash:

--ExpireInfoBar, removes delegate from infobar_delegates_
--WebContents::DidStopLoading, causes PasswordManager to call AddInfoBar because it
has a pending_save_manager_ (and AddInfoBar succeeds because there is no "equal" 
delegate already in the set)
--Animation terminates, which calls InfoBarClosed(), and resets the 
pending_save_manager_
--Now we have an infobar and a null pending_save_manager_, so clicking Accept/Cancel 
will deref 0.

Renamed |pending_save_manager_| to |provisional_save_manager_| to be more consistent,
and added |pending_decision_manager_| to take responsibility of the PasswordFormManager
in |provisional_save_manager_| when the infobar comes in to play, so that we have
the same ownership model as before the new infobar changes.


Review URL: http://codereview.chromium.org/14464

git-svn-id: http://src.chromium.org/svn/trunk/src@7094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c