Make the linter just compare base paths instead of looking at all the files on disk. Now it takes ~3 seconds on my machine.
Also, make sure that each path in the test lists actually corresponds to a file/directory on disk.  This is slightly stricter than we need to be as we could do partial paths, but I don't see the benefit in supporting that. As it is, I've found a lot of tests that are listed incorrectly in the test_lists because they point to paths that don't exist.
Review URL: http://codereview.chromium.org/14099

git-svn-id: http://src.chromium.org/svn/trunk/src@7154 4ff67af0-8c30-449e-8e8b-ad334ec8d88c