Rebaseline numeric-conversions per ager's advice.
The expected files expect a FAIL line because I can't change the upstream test. Copying the test to the platform/directories doesn't help because the test is actually in a resources/xxx.js file and there seem to be problems with paths in that case.

Review URL: http://codereview.chromium.org/15201


git-svn-id: http://src.chromium.org/svn/trunk/src@7163 4ff67af0-8c30-449e-8e8b-ad334ec8d88c