Reverting 7423.
After talking with Darin, I'm convinced this was a bad idea.  The issue is that Linux can't actually know the encoding of the file (bad Linux!), so on Linux this method is nonsensical.  It also turns out to not be useful for the case I thought I needed it for.
Review URL: http://codereview.chromium.org/16252

git-svn-id: http://src.chromium.org/svn/trunk/src@7455 4ff67af0-8c30-449e-8e8b-ad334ec8d88c