Add chrome-user-script:// protocol
This is a step towards getting user scripts working in extensions. It's a bit
janky to use the form chrome-user-script://<script file>/ (with no path), but
GURL assumes that there is always a host, but path is optional, making this
approach simpler than alternatives.

Review URL: http://codereview.chromium.org/16592

git-svn-id: http://src.chromium.org/svn/trunk/src@7759 4ff67af0-8c30-449e-8e8b-ad334ec8d88c