Move webkit_resources.rc to webkit_resources.grd.  Add webkit_resources.vcproj which creates the .rc files and have webkit glue depend on it.
I had to add a flag to grit that allows a resource to be in the
header, but not in the .rc file.  We do this for a test_shell only
resource that is not compiled in.

SCons build also works on windows.

I still need to update the other .sln files and mac build.

Review URL: http://codereview.chromium.org/17466

git-svn-id: http://src.chromium.org/svn/trunk/src@7910 4ff67af0-8c30-449e-8e8b-ad334ec8d88c