Fix a bunch of svg tests for text rendering.
There were 18 tests in the svg/text and svg/W3C-SVG-1.1/text* that were 
failing for a number of reasons:

1) Selection color is darker on Windows than Mac, so we expect dark blue 
boxes around the text whereas Mac expects light-blue. Why does this matter?
Because most of the tests in the svg/text folder perform a Select All 
at the end of the test.

2) Changes in WebKit where handling of font-weight was delegated to Windows, 
which affects font-weight="600" in two tests (600 looks narrower).

3) Some tests simply needed expected text output.

4) And of course, the usual missing Z from the metric output.


Review URL: http://codereview.chromium.org/18374

git-svn-id: http://src.chromium.org/svn/trunk/src@8314 4ff67af0-8c30-449e-8e8b-ad334ec8d88c