Fix a bug when saving web pages that have a period in their title.
The save dialog code attempts to generate a file extension based on
the suggested name provided to it, which in the case of saving a web
page is the title. If that title contains a '.' character, it treats
anything that follows as the file extension and appends it to the name
the user selects.

For example, saving a page with the title "Google Inc. - A search engine"
and the user choosing the save name as "Google" will result in the saved
file being named "Google. - A search engine.htm" when it should be
"Google.htm".

In the case of saving a web page, we can ignore trying to figure out
the file extension since we already know it will be ".htm".

BUG=6105 (http://crbug.com/6105)

Review URL: http://codereview.chromium.org/18699

git-svn-id: http://src.chromium.org/svn/trunk/src@8569 4ff67af0-8c30-449e-8e8b-ad334ec8d88c