Try again to commit r8486, which updates the manifest parsing in the Extension class to also parse out user script details.
The problem was that extension_protocols.cc appears to
link OK if you don't refer to anything in it, but as
soon as you try to use something defined in it, a bunch
of other symbols it refers to turn up undefined.

To avoid this problem in the future, I stopped compiling
extensions_protocol.cc so that we would realize if we
tried to use it.

Also in this change, I absolutified paths in
ExtensionsService before creating Extension instances.
On Linux, the path that PathService was giving us was
not absolute.

Review URL: http://codereview.chromium.org/18704

git-svn-id: http://src.chromium.org/svn/trunk/src@8576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c