This change should have few user-visible effects.  I think I slightly modified the window layout (by a pixel or so) in the popup window case, which I haven't tried to make pixel-perfect yet anyway.  Otherwise it's all reorg.
Clean up logic of nonclient view to be more consistent and coherent.  Most things are now based on a small set of core functions that return various frame/nonclient thicknesses.  This allows a noticeable reduction in the number of hardcoded constants, makes the code make more sense (even though a few places are longer than before), and makes it much easier to do things in the future like resizing various borders or handling system metrics better.

Rename and update comments for all constants to be more descriptive and less redundant.

Rename helper functions and variables for consistency, now that they've changed.

Remove an unused variant of TileImageInt().  Add a new one that takes an offset into the source image.  This is used to rewrite the maximized bottom border tiling call in a way that doesn't rely on someone else clipping out or painting over the overdraw from tiling "too large" a source image.  Probably no visible effect.

Make resize border hittest helper function able to handle borders of varying widths, and try to make parameter names more meaningful.  Technically, we don't need the extra flexibility, since our bottom and side borders are the same thickness, but since there are different functions calculating that thickness on the caller side, it didn't hurt to not rely on that for the future.

BUG=5054
Review URL: http://codereview.chromium.org/18804

git-svn-id: http://src.chromium.org/svn/trunk/src@8658 4ff67af0-8c30-449e-8e8b-ad334ec8d88c