Generate a proper file name when dragging an image from a web page to the desktop.
We receive information about the image from WebKit in the
render which includes the image URL, MIME type and a guess
as to what the file name is. If the filename is empty, we
synthesize a file name based on the URL. In all cases, we
set the file extension to the MIME type that was detected
by WebKit.

This change is the Chromium portion of image dragging. The
WebKit portion has been committed as:
http://trac.webkit.org/changeset/40294

BUG=6481 (http://crbug.com/6481)

Review URL: http://codereview.chromium.org/19417

git-svn-id: http://src.chromium.org/svn/trunk/src@8854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c