Remove old web app code it's no longer needed. Simplifies startup a little.
Also fixes a regression from the BrowserInit porting work. I was a little over-zealous removing the CommandLine handling. It turns out that it's this that MessageWindow uses to convey the command line of the signalling process to the existing process, and that the existing process re-runs BrowserInit with the signalling process' command line. So I plumbed that back through again.

Review URL: http://codereview.chromium.org/19746

git-svn-id: http://src.chromium.org/svn/trunk/src@9034 4ff67af0-8c30-449e-8e8b-ad334ec8d88c