Fix a crash where we would try to add a DocumentMarker with the same start- and endOffset, which AddMarker in WebKit will ignore. As a result, our marker vector is empty and we try to access the element with index -1.
Also added a check to avoid a NOTREACHED in debug
build, which can happen under normal circumstances 
(when you navigate away from a page when a search 
is in progress).

BUG=http://crbug.com/7250
TEST=We don't have reproduction steps for the
crash but if we stop seeing this on our crash
server we'll know that it is fixed.
Review URL: http://codereview.chromium.org/20028

git-svn-id: http://src.chromium.org/svn/trunk/src@9105 4ff67af0-8c30-449e-8e8b-ad334ec8d88c