Fix for http://code.google.com/p/chromium/issues/detail?id=7429
The problem is that webkit's timers are all based off the WTF/CurrentTime
implementation, but the bridge was doing math between a webkit that
implementation of currentTime and Chromium's base/Time implementation.
Subtracting two different times leads to skew, which turned out to
always be negative, so the CPU was spinning for most all timers.  

This bug was introduced during the merge which brought in timers.  From
the outside, functionality was correct (timers fired at the right time),
but internally we could spin the CPU in while waiting for the timer to
fire.

There is some code to remove from SystemTimeChromium, which I will do,
but that is in the webkit tree, so I'll do that separately.


Review URL: http://codereview.chromium.org/21122

git-svn-id: http://src.chromium.org/svn/trunk/src@9307 4ff67af0-8c30-449e-8e8b-ad334ec8d88c