Make the mime -> extension code return a dotless extension.
Previously we would return something like ".html", because that's how it's stored in the registry on Windows.  For a variety of reasons, it's much simpler to work without the prefixed dot.  In fact, the only two consumers were manually stripping of the prefix dot.

There is a cooresponding change that should happen in WebKit / ChromiumBridge, but since it's checking if there is a prefix dot and removing it, the code will still work, and I'll follow up with that change.

Review URL: http://codereview.chromium.org/23010


git-svn-id: http://src.chromium.org/svn/trunk/src@9333 4ff67af0-8c30-449e-8e8b-ad334ec8d88c