Linux: fix failure to find any fonts
When we are looking for a fallback font, we don't want the usual
family name matching to stop us from finding one. The only case where
we should fail to find a fallback is when fontconfig doesn't know
about /any/ fonts on the system.

This patch adds a flag argument to FontMatch which tells it when we're
in fallback mode.


git-svn-id: http://src.chromium.org/svn/trunk/src@9477 4ff67af0-8c30-449e-8e8b-ad334ec8d88c