Fixing two layout tests.
One new test (fast/runin/002.html) was added and 
needed a baseline since the only difference is in
font metrics.

Expectations for fast/inline/percentage-margins 
changed upstream. The change log just says it is 
rebaselined "due to recent changes", but when I 
compare the changes they are identical (left side 
moves 4 pixels to the left and the right moves 8 
pixels to the left).

Review URL: http://codereview.chromium.org/21219

git-svn-id: http://src.chromium.org/svn/trunk/src@9502 4ff67af0-8c30-449e-8e8b-ad334ec8d88c