Maintain a refcounted AutomationProvider pointer in the ExternalTabContainer. This ensures that we don't crash the browser while trying to dereference a freed AutomationProvider pointer.
When a Chrome browser instance starts up, we attempt to locate an already running instance and defer to it to complete the navigation request. However it is quite possible for the running instance to exit while we attempt to send a WM_COPYDATA message to it. This caused a bunch
of ASSERTS to fire off in the browser. 

Fixes as below:-
1. If GetWindowThreadProcessId fails, we bail out and try to launch a new chrome instance
2. If SendMessageTimeout fails, we check if the window is still valid. If not we bail out and try to launch a new chrome instance.
3. We return an error from the WM_COPYDATA handler if the chrome instance is in the process of shutting down. We handle this at the caller end, i.e. in NotifyOtherProcess.

Bug=1643310

Review URL: http://codereview.chromium.org/23016

git-svn-id: http://src.chromium.org/svn/trunk/src@9536 4ff67af0-8c30-449e-8e8b-ad334ec8d88c