Fetch SDCH dictionary as soon as current URL fetch completes
Rather than just waiting for 15 seconds, this patch fetches the
"suggested dicitionary" as soon as the current download completes.
This avoids stealing bandwidth from the original download, but 
works as quickly as possible to get the dictionary, so that the
user can receive compression results asap.

r=wtc
Review URL: http://codereview.chromium.org/20254

git-svn-id: http://src.chromium.org/svn/trunk/src@9597 4ff67af0-8c30-449e-8e8b-ad334ec8d88c