The DOM Automation controller object uses the RenderView instance as the message sender, i.e. to send replies to javascript requests issued by the browser. The DOM automation controller object is bound to the frame in the WindowObjectcleared code path.The current implementation maintains the message sender object as a static pointer, which causes a crash if the RenderView instance goes out of scope.
This can be reproduced with a page which causes a popup window to show up and close. If we attempt to use the Dom Automation controller instance bound to the original Renderview, which is still valid, we crash.

The fix is to maintain the message sender as a member variable. The lifetime of the Dom Automation controller instance depends on the RenderView lifetime anyway as it is a member variable. This mimics the other CppBindings like the external host bindings, etc.

Added an automation test to test this case. I verified that the test crashes without this fix.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=3941

Bug=3941

Review URL: http://codereview.chromium.org/21441

git-svn-id: http://src.chromium.org/svn/trunk/src@9963 4ff67af0-8c30-449e-8e8b-ad334ec8d88c